create policy set file
